==> src/4. Boruta Feature Selection.Rmd <==
---
title: "Tropical Fever Project"
author: "Victor Yman & Zaynab Mousavian"
date: "Dec. 2023"
output:
  html_document: default
  pdf_document: default
subtitle: Boruta Feature Selection
---

### R packages setup and loading data

```{r warning = FALSE, message = FALSE }
library(Boruta)
library(tidyverse)
library(RColorBrewer)
l=load("../Data/ReadyData.RData") 
types=data.frame(Type=setdiff(levels(Group),c("Healthy control","Pulmonary bacteria")),Color=brewer.pal(8, "Set1")[c(1,2,4,5)],Label=c("A","B","C","D"))
row.names(cyt_log)=gsub("\\.","",row.names(cyt_log))
```

### Excluding healthy controls

```{r}
cyt_log = cyt_log[,which(Group!="Healthy control")]
Group = Group[which(Group!="Healthy control")]
```

### Adding clinical variables

```{r}
library(readxl)
clinical_data <- read_excel("../Data/TF-joint clinical table_selected_stats_vJun22.xlsx")
clinical_data <- clinical_data %>% filter(ID%in%colnames(cyt_log))
row.names(clinical_data) <- clinical_data$ID
clinical_data <- clinical_data %>% arrange(match(ID,colnames(cyt_log)))
clinical_data <- clinical_data %>% dplyr::select(Age,Maxtemp,P_CRP,B_WBC,B_PLT)
clinical_data <- t(clinical_data)
merged_data <- rbind(cyt_log,clinical_data)
features <- row.names(merged_data)
merged_data <- as_tibble(merged_data)
merged_data <- merged_data %>%
  select_if(~all(!is.na(.)))
Group <- Group[colnames(cyt_log)%in%colnames(merged_data)]
cyt_log <- as.data.frame(merged_data)
row.names(cyt_log) <- features
```

### Run Boruta (Figures 4 and S7 and S8)

```{r message=FALSE, warning=FALSE}
candidates <- data.frame()
Allcandidates <- data.frame()
pdf(file = paste0("../Results/Boruta/ALL_boruta.pdf"),height = 20,width = 10)
layout(matrix(c(1,2,3,4),ncol=1), widths=c(4,4,4,4), heights=c(1.5,1.5,1.5,1.5), TRUE)
runBoruta <- function(x){
  type=x[1]
  color=x[2]
  label=x[3]
  cyt_data <- as.data.frame(t(cyt_log))
  cyt_data$Group=factor(Group)
  newVar <- rep(type, nrow(cyt_data))
  newVar[which(cyt_data$Group!=type)] = paste0("No ",type)
  newVar <- as.factor(newVar)
  cyt_data$Group=newVar
  set.seed(111)
  boruta_data <- Boruta(Group~., data = cyt_data, doTrace = 2, maxRuns=400)
  boruta_data <- TentativeRoughFix(boruta_data)
  boruta_stats <- attStats(boruta_data)
  boruta_stats$cytokine <- rownames(boruta_stats)
  boruta_stats <- boruta_stats %>% filter(decision =="Confirmed")
  boruta_stats_New <- boruta_stats[order(boruta_stats$meanImp,decreasing = TRUE),c("cytokine","meanImp")]
  Allcandidates <<- rbind(Allcandidates, cbind(boruta_stats_New,infection=type))
  
  Thr <- (max(boruta_stats$meanImp)-min(boruta_stats$meanImp))/2+min(boruta_stats$meanImp)
  boruta_stats <- boruta_stats %>% filter(meanImp > Thr )
  boruta_stats <- boruta_stats[order(boruta_stats$meanImp,decreasing = TRUE),c("cytokine","meanImp")]
  candidates <<- rbind(candidates, cbind(boruta_stats,infection=type))
  plot(boruta_data, cex.axis = 1.1, xlab = "", las = 2,ylab="")
  title(label,line=+1,adj=0,cex.main=2)
}
l=apply(types,1,function(x) runBoruta(x))
dev.off()
write.table(Allcandidates, paste0("../Results/Boruta/Allcandidates.txt"), row.names=FALSE, quote=FALSE,sep=",")

write.table(candidates, paste0("../Results/Boruta/candidates.txt"), row.names=FALSE, quote=FALSE,sep=",")
maxImp <- aggregate(meanImp ~ cytokine, data = candidates, max)
maxImp <- merge(maxImp,candidates)
write.table(maxImp, paste0("../Results/Boruta/maxImpCandidates.txt"), row.names=FALSE, quote=FALSE,sep=",")
```

### Show selected cytokines for different pathogens (Figure S6)

```{r}
candidates <- read.delim("../Results/Boruta/Allcandidates.txt",sep=",")
path_cols <- brewer.pal(8, "Set1")[c(1,2,4,5)]
library(ggplot2)
pdf(file="../Results/Boruta/SelectedCytokines.pdf",width=10,height = 4)
ggplot(candidates,aes(x=cytokine,y=infection,color=infection))+
  geom_point(size=3)+
  theme_bw()+
  labs(x="Cytokines",y="Pathogens")+
  theme(text=element_text(size=12),axis.title = element_text(face=2),axis.text.x=element_text(angle=90,hjust=1,vjust=0.5,size=12,face=1),axis.text.y=element_text(vjust=0.5,hjust=1,size=12,face=1),legend.title = element_text(size=10,face=2))+
  scale_color_manual(values=path_cols)+
  guides(color=FALSE)
dev.off()
```
