==> src/5. ROC Analysis.Rmd <==
---
title: "Tropical Fever Project"
author: "Victor Yman & Zaynab Mousavian"
date: "Dec. 2023"
output:
  html_document: default
  pdf_document: default
subtitle: ROC Analysis
---

### R packages setup and loading data

```{r message=FALSE, warning=FALSE}
library(pROC)
library(caret)
library(groupdata2)
library(kernlab)
library(Hmisc)
library(tidyverse)
library(ROCaggregator)
library(RColorBrewer)
library(ROCR)
library(pracma)
library(stats)
library(caretEnsemble)
library(PRROC)
plots=list()
plots2=list()
```

### Loading data for classification of each pathogen

```{r message=FALSE, warning=FALSE}
l=load("../Data/ReadyData.RData")
i=1
type2=setdiff(levels(Group),c("Healthy control","Pulmonary bacteria"))[i]
if (i==1)
  type="DengueVirus"
if (i==2)
  type="EntericBacteria"
if (i==3)
  type="InfluenzaVirus"
if (i==4)
  type="Pfalciparum"



candidates=read.delim("../Results/NewBoruta2/maxImpCandidates.txt",sep=",",header=TRUE)
candidates=read.delim("../Results/NewBoruta2/candidates.txt",sep=",",header=TRUE)
candidates=read.delim("../Results/NewBoruta2/Allcandidates.txt",sep=",",header=TRUE)
cytokines=candidates[which(candidates$infection==type2),"cytokine"]
cytokines
```

### Excluding healthy controls

```{r}
cyt_log = cyt_log[,which(NewGroup!="HealthyControl")]
NewGroup = NewGroup[which(NewGroup!="HealthyControl")]
```

### Adding clinical variables

```{r}
library(readxl)
clinical_data <- read_excel("../Data/TF-joint clinical table_selected_stats_vJun22.xlsx")
clinical_data <- clinical_data %>% filter(ID%in%colnames(cyt_log))
row.names(clinical_data) <- clinical_data$ID
clinical_data <- clinical_data %>% arrange(match(ID,colnames(cyt_log)))
clinical_data <- clinical_data %>% dplyr::select(Age,Maxtemp,P_CRP,B_WBC,B_PLT)
clinical_data <- t(clinical_data)
merged_data <- rbind(cyt_log,clinical_data)
features <- row.names(merged_data)
merged_data <- as_tibble(merged_data)
merged_data <- merged_data %>%
  select_if(~all(!is.na(.)))
NewGroup <- NewGroup[colnames(cyt_log)%in%colnames(merged_data)]
cyt_log <- as.data.frame(merged_data)
row.names(cyt_log) <- features
```

### Make data ready for ROC analysis

```{r message=FALSE, warning=FALSE}
NewGroup= as.data.frame(NewGroup)
row.names(cyt_log) <- gsub("\\.", "", row.names(cyt_log))
row.names(NewGroup)=colnames(cyt_log[cytokines,,drop=FALSE])
data_Selected = cyt_log[cytokines,,drop=FALSE]
data = as.data.frame(t(data_Selected))
data$Class = factor(NewGroup$NewGroup)
data$Sample=as.factor(row.names(data))
```

### Run random forest classification by 5-fold cross validation and compute ROC for each fold

```{r}
set.seed(6)
set.seed(10)
data <- fold(data, k = 5, cat_col = "Class", id_col = 'Sample')

newVar <- rep(type, ncol(cyt_log))
newVar[which(NewGroup$NewGroup!=type)] = "Control"
newVar <- as.factor(newVar)
NewGroup$NewGroup=newVar
data$Class = factor(NewGroup$NewGroup)

rocs <- list()
confmat <- list()
counts <- list()
prs=list()
# One iteration per fold
for (fold in 1:5){
  
  # Create training set for this iteration
  # Subset all the datapoints where .folds does not match the current fold
  training_set <- data[data$.folds != fold,]
  
  # Create test set for this iteration
  # Subset all the datapoints where .folds matches the current fold
  testing_set <- data[data$.folds == fold,]
  training=subset(as.data.frame(training_set),select=-c(Sample,.folds))
  testing=subset(as.data.frame(testing_set),select=-c(Sample,.folds))
  
  set.seed(10)
  set.seed(15)
  fitControl <- trainControl(method = "repeatedcv",
                             ## 10-fold CV...
                             repeats = 10,
                             number = 10,
                             classProbs = TRUE,
                             summaryFunction = defaultSummary)
  
   model <- train(Class ~ ., data = training, 
                 method = "ranger", 
                 trControl = fitControl,metric="Kappa")
  
  
  pred <- predict(model, testing)
  prob <- predict(model, testing, type="prob")
  mat0 <- confusionMatrix(pred,testing$Class)
  
  roc0 <- roc(response = testing$Class, 
              predictor = prob[,1], 
              levels = rev(levels(testing$Class)),ci=TRUE)
  pr<-pr.curve(scores.class0 = testing$Class, weights.class0 = prob[,1],curve=TRUE);
  
  rocs[[paste0("fold:",fold)]]=roc0
  prs[[paste0("fold:",fold)]]=pr
  confmat[[paste0("fold:",fold)]]=mat0
  
  counts[[paste0("fold:",fold)]]=c(sum(testing$Class == "Control"),nrow(testing))
}
pdf("test.pdf")
plot(prs[[1]])
dev.off()
```

### Aggregate ROC and PR plots of different folds into one main plot 

```{r}
get_proc <- function(roc_obj,counts_obj){
  list(
    "fpr" = 1 - roc_obj$specificities,
    "tpr" = roc_obj$sensitivities,
    "thresholds" = roc_obj$thresholds,
    "negative_count"= counts_obj[1],
    "total_count" = counts_obj[2],
    "auc" = roc_obj$auc
  )
}

roc_obj_node_1 <- get_proc(rocs[[1]],counts[[1]])
#> Setting levels: control = 0, case = 1
#> Setting direction: controls < cases
roc_obj_node_2 <- get_proc(rocs[[2]],counts[[2]])
#> Setting levels: control = 0, case = 1
#> Setting direction: controls < cases
roc_obj_node_3 <- get_proc(rocs[[3]],counts[[3]])
#> Setting levels: control = 0, case = 1
#> Setting direction: controls < cases
roc_obj_node_4 <- get_proc(rocs[[4]],counts[[4]])
#> Setting levels: control = 0, case = 1
#> Setting direction: controls < cases
roc_obj_node_5 <- get_proc(rocs[[5]],counts[[5]])
#> Setting levels: control = 0, case = 1
#> Setting direction: controls < cases

# Preparing the input
fpr <- list(roc_obj_node_1$fpr, roc_obj_node_2$fpr, roc_obj_node_3$fpr,roc_obj_node_4$fpr,roc_obj_node_5$fpr)
tpr <- list(roc_obj_node_1$tpr, roc_obj_node_2$tpr, roc_obj_node_3$tpr,roc_obj_node_4$tpr,roc_obj_node_5$tpr)
thresholds <- list(
  roc_obj_node_1$thresholds, roc_obj_node_2$thresholds, roc_obj_node_3$thresholds,roc_obj_node_4$thresholds,roc_obj_node_5$thresholds)
negative_count <- c(
  roc_obj_node_1$negative_count, roc_obj_node_2$negative_count, roc_obj_node_3$negative_count,roc_obj_node_4$negative_count,roc_obj_node_5$negative_count)
total_count <- c(
  roc_obj_node_1$total_count, roc_obj_node_2$total_count, roc_obj_node_3$total_count,roc_obj_node_4$total_count,roc_obj_node_5$total_count)

# Compute the global ROC curve for the model
roc_aggregated <- roc_curve(fpr, tpr, thresholds, negative_count, total_count)

roc_aggregated$sensitivities <- roc_aggregated$tpr
roc_aggregated$specificities <- 1-roc_aggregated$fpr
# Calculate the AUC
roc_auc <- trapz(roc_aggregated$fpr, roc_aggregated$tpr)

sprintf("ROC AUC aggregated from each node's results: %f", roc_auc)

precision_recall_aggregated <- precision_recall_curve(
  fpr, tpr, thresholds, negative_count, total_count)
# Calculate the precision-recall AUC
precision_recall_auc <- -trapz(
  precision_recall_aggregated$recall, precision_recall_aggregated$pre)

sprintf(
  "Precision-Recall AUC aggregated from each node's results: %f",
  precision_recall_auc
)
```

### Visualize results using ROC plot

```{r}
path_cols <- brewer.pal(8, "Set1")[c(1,2,4,5)]

new_rocs=data.frame()
j=1
roc_folds <- function(roc_obj){
  new_rocs <<- rbind(new_rocs,data.frame(tpr=roc_obj$sensitivities,fpr=1-roc_obj$specificities,fold=paste0("Fold ",as.character(j)),color="gray"))
  new_rocs <<- new_rocs[order(new_rocs$fold,new_rocs$fpr,new_rocs$tpr,decreasing = FALSE),]
  j <<- j+1
}
lapply(rocs,function(x) roc_folds(x))
rocPlot<- data.frame(tpr=roc_aggregated$tpr,fpr=roc_aggregated$fpr,fold="Average of Folds",color="notGray")

plots[[i]] <- ggplot(new_rocs,aes(x = fpr, y = tpr,group=fold))+
  geom_line(aes(linetype=fold),colour='gray')+
  geom_line(data=rocPlot, aes(x = fpr, y = tpr),linetype=1,color=path_cols[i],linewidth=2)+
  labs(x = "1 - Specificity", y = "Sensitivity")+
  ggtitle(paste0("Average AUC=",format(round(roc_auc,2),nsmall=2)))+
  scale_linetype_manual(values=c(2,3,4,5,6))+
  theme_bw()+
  theme(plot.title = element_text(size=14, face="bold",color=path_cols[i]),axis.title = element_text(size=14,face="bold"),axis.text = element_text(size=14))+
  guides(linetype=FALSE)

```

### Visualize results using Precision-recall plot

```{r}
path_cols <- brewer.pal(8, "Set1")[c(1,2,4,5)]


prPlot<- data.frame(pre=rev(precision_recall_aggregated$pre),tpr=rev(precision_recall_aggregated$recall),fold="Average of Folds",color="notGray")

plots2[[i]] <- ggplot(data=prPlot, aes(x = tpr, y = pre))+
  geom_line(linetype=1,color=path_cols[i],linewidth=2)+
  labs(x = "Recall", y = "Precision")+
  ggtitle(paste0("Average AUC=",format(round(precision_recall_auc,2),nsmall=2)))+
  theme_bw()+
  theme(plot.title = element_text(size=14, face="bold",color=path_cols[i]),axis.title = element_text(size=14,face="bold"),axis.text = element_text(size=14))+
  guides(linetype=FALSE)

```

### Plot all ROC plots in one plot (Figures 5, S7 and S8)

```{r}
save(plots,file="../Results/ROC/ALL-ROCs-all cytokines.RData")
save(plots2,file="../Results/ROC/ALL-PR-all cytokines.RData")

pdf(file = paste0("../Results/ROC/ALL_ROCs-all cytokines.pdf"))
plot_grid(plots[[1]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)), plots[[2]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)),plots[[3]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)),plots[[4]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)))
dev.off()

pdf(file = paste0("../Results/ROC/ALL_PR-all cytokines.pdf"))
plot_grid(plots2[[1]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)), plots2[[2]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)),plots2[[3]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)),plots2[[4]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)), labels = c('A', 'B','C','D'), label_size = 14)
dev.off()

pdf(file = paste0("../Results/ROC/All plots.pdf"),width = 16,height = 8)
plot_grid(plots[[1]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)), plots[[2]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)),plots[[3]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)),plots[[4]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)),plots2[[1]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)), plots2[[2]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)),plots2[[3]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)),plots2[[4]]+theme(axis.title = element_text(face=1,size=10),axis.text = element_text(size=10)), ncol = 4,nrow=2)
dev.off()
```
